check and correct tests, correct logic
